requirements.txt
streamlit>=1.30
pandas>=2.0
plotly>=5.18
python-pptx>=0.6.21
anthropic>=0.40
python-dotenv>=1.0
openpyxl>=3.1
kaleido>=0.2.1
markdown>=3.4
